changed default config directory to be ./.rover
it was 'config' and based on the config file's location.
this won't work in an environment where the code needs to be deployed
to an unwritable location.  by prefixing w/ '.', it will be hidden and
can fit more nicely closer in a root directory.  the change to 'rover'
makes it obvious what's in the directory rather than just random "config"